surgical-liquid-S5/requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
art==5.8
async-timeout==4.0.1
attrs==21.2.0
av==10.0.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blinker==1.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
chex==0.1.6
click==8.0.3
cloudpickle==1.6.0
cmake==3.26.3
ConfigArgParse==1.5.3
cryptography==3.4.8
cycler==0.10.0
datasets==2.12.0
decorator==4.4.2
dill==0.3.6
dm-control==1.0.9
dm-env==1.6
dm-tree==0.1.8
docker-pycreds==0.4.0
einops==0.3.0
etils==1.3.0
filelock==3.10.0
flatbuffers==2.0.7
flax==0.6.8
frozenlist==1.2.0
fsspec==2023.3.0
future==0.18.3
gast==0.5.4
gitdb==4.0.10
GitPython==3.1.31
glfw==2.5.9
google-auth==1.35.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.59.0
grpcio==1.42.0
gym==0.19.0
huggingface-hub==0.13.2
idna==3.1
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.8.2
importlib-resources==5.12.0
iopath==0.1.10
ipdb==0.13.9
ipython==7.27.0
ivtmetrics==0.1.0
jax==0.4.7
#jaxlib==0.4.7+cuda11.cudnn82 -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jedi==0.18.0
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.3.2
labmaze==1.0.6
lazy_loader==0.1
lightning-utilities==0.7.1
lit==16.0.5
lxml==4.9.2
Markdown==3.3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.4.1
matplotlib-inline==0.1.3
mdurl==0.1.2
ml-dtypes==0.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.5
mujoco==2.3.5
multidict==5.2.0
multiprocess==0.70.14
ncps==0.0.7
nest-asyncio==1.5.6
networkx==2.8.6
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.1
olefile==0.46
omegaconf==2.3.0
opencv-python==4.5.5.64
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
optax==0.1.5
orbax==0.1.7
packaging==23.0
pandas==1.2.4
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
portalocker==2.7.0
proglog==0.1.10
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycm==3.8
pycparser==2.21
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.3.0
PyOpenGL==3.1.7
pyOpenSSL==21.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.9.4
pytz==2021.3
pyu2f==0.1.5
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
regex==2023.3.23
requests==2.26.0
requests-oauthlib==1.3.0
responses==0.18.0
rich==13.4.1
rsa==4.8
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.7
S5==0.1
scikit-image==0.20.0
scikit-learn==1.1.2
scipy==1.9.1
sentry-sdk==1.16.0
setproctitle==1.3.2
setuptools==58.0.4
simplejson==3.18.4
six==1.16.0
smmap==5.0.0
sympy==1.12
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow-datasets==4.5.2
tensorflow-metadata==1.13.1
tensorflow-probability==0.20.1
tensorstore==0.1.37
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2023.2.28
timm==0.6.12
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torch-tb-profiler==0.3.1
torchdata==0.6.1
torchmetrics==0.11.3
torchsummary==1.5.1
torchtext==0.15.2
torchvision==0.15.2
tqdm==4.62.3
traitlets==5.1.0
transformers==4.28.0
triton==2.0.0
typing_extensions==4.6.3
urllib3==1.26.14
wandb==0.13.11
wcwidth==0.2.5
Werkzeug==2.0.1
wheel==0.37.0
xxhash==3.2.0
yacs==0.1.8
yarl==1.7.2
zipp==3.6.0
